// client/src/app/_models/hmi.ts
export class GaugeSettings {
    name = '';
    property: any = null;   // set to GaugeProperty after upgrate
    label = '';             // Gauge type label
    hide = false;
    lock = false;
    constructor(public id: string, public type: string) {
    }
}
// ...
export enum GaugeActionsType {
    hide = 'shapes.action-hide',
    show = 'shapes.action-show',
    blink = 'shapes.action-blink',
    color = 'shapes.action-color',
    stop = 'shapes.action-stop',
    clockwise = 'shapes.action-clockwise',
    anticlockwise = 'shapes.action-anticlockwise',
    downup = 'shapes.action-downup',
    rotate = 'shapes.action-rotate',
    move = 'shapes.action-move',
    monitor = 'shapes.action-monitor',
    refreshImage = 'shapes.action-refreshImage',
    start = 'shapes.action-start',
    pause = 'shapes.action-pause',
    reset = 'shapes.action-reset',
    enable = 'shapes.action-enable',
    disable = 'shapes.action-disable',
}

export class GaugeAction {
    variableId: string;
    bitmask: number;
    range: GaugeRangeProperty;
    type: any;
    options: any = {};
}
// ...
export class GaugeStatus {
    variablesValue = {};
    onlyChange = false;         // to process value only by change
    takeValue = false;          // to process value by check change with gauge value
    actionRef: GaugeActionStatus;
    disabled = false;           // button disabled state (blocks event execution)
    activeHiders?: {[key: string]: boolean};  // tracks which hide actions are currently active
}

export class GaugeActionStatus {
    type: string;
    timer?: any = null;
    animr?: any = null;
    spool?: any;
    constructor(type: string) {
        this.type = type;
    }
}
// ...
export class GaugeRangeProperty {
    min: number;
    max: number;
    minVariableId?: string;
    maxVariableId?: string;
    text: string;
    textId: string;
    color: string;
    type: any;
    style: any;
    stroke: string;
    address?: string;
}
// ...
export class Variable {
    id: string;
    name: string;
    source: string;
    value: string;
    error: number;
    timestamp: number;
    device?: Device;
    quality?: string;
    constructor(id: string, name: string, device?: Device) {
        this.id = id;
        this.name = name;
        this.device = device;
        if (device?.type === DeviceType.internal) {
            this.value = '0';
        }
    }
}